- Rename JDK8 package to webservices.jdk8 - Add Initial Webservice JDK11 package - Add pom/config.xml files to JDK11 package - Added initial component for new JDK 11 WebServiceDriver HttpClient
