+++ Framework/jmaqs-webservices-jdk11/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <groupId>com.magenic.jmaqs</groupId>
+        <artifactId>jmaqs-framework</artifactId>
+        <version>0.0.1-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <groupId>com.magenic.jmaqs.webservices</groupId>
+    <artifactId>jmaqs-webservices-jdk11</artifactId>
+    <name>jmaqs-webservices-jdk11</name>
+    <packaging>jar</packaging>
+    <url>http://maven.apache.org</url>
+
+
+</project>
