--- Framework/jmaqs-webservices/src/test/java/com/magenic/jmaqs/webservices/models/Product.java
@@ -1,70 +1,0 @@
-package com.magenic.jmaqs.webservices.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.math.BigDecimal;
-
-public class Product {
-    @JsonProperty("Id")
-    private int id;
-
-    @JsonProperty("Name")
-    private String name;
-
-    @JsonProperty("Category")
-    private String category;
-
-    @JsonProperty("Price")
-    private BigDecimal price;
-
-    public Product(int id, String name, String category, BigDecimal price) {
-        this.id = id;
-        this.name = name;
-        this.category = category;
-        this.price = price;
-    }
-
-    public Product() { }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getCategory() {
-        return category;
-    }
-
-    public void setCategory(String category) {
-        this.category = category;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-
-    public String toString() {
-        StringBuilder sb = new StringBuilder(20);
-        sb.append(String.format("%s:%d\n", "Id", this.getId()));
-        sb.append(String.format("%s:%s\n", "Name", this.getName()));
-        sb.append(String.format("%s:%s\n", "Category", this.getCategory()));
-        sb.append(String.format("%s:%d\n", "Price", this.getPrice()));
-
-        return sb.toString();
-    }
-}
